+++ react-app/src/routes/PlayPong.tsx
@@ -2,7 +2,7 @@ export default function PlayPong() {
   return (
     <>
       <div>
-        <h1 className="text-rose-700Í text-2xl">
+        <h1 className="text-rose-50 text-4xl bg-gradient-to-tl from-rose-400 to-sky-400 p-10 rounded-full px-20 shadow-2xl font-bold ">
           Wanna Play Some Pong?
         </h1>
       </div>
